requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
empath==0.89
idna==3.10
joblib==1.4.2
lxml==5.3.0
nltk==3.9.1
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0